+++ CHANGELOG.md
@@ -47,6 +47,7 @@ All notable changes to this project will be documented in this file.
 - *(sync)* Added some sync process with api
 - *(sync)* Added some sync process with api for pdf
 - *(unkown)* I forgot but is good to save
+- *(ui)* Improved cards design at boarding screen
 
 ### 🐛 Bug Fixes